my_discord/requirements.txt
discord.py
python-dotenv
flask
requests
mysql-connector-python